backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator  # Required for pydantic[email]

# Google Cloud
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
google-cloud-documentai>=2.20.0  # Document AI Python 클라이언트
google-cloud-firestore
google-auth>=2.0.0
google-auth-oauthlib
requests

# NLP & ML (these are not currently used in the code)
# transformers==4.38.0  # Heavy, not used
# sentence-transformers==2.2.2  # Heavy, not used
# konlpy==0.6.0  # Optional: Requires Java
# scikit-learn==1.3.0  # Not used in core API
numpy==1.24.0
# pandas==2.1.0  # Not used in core API

# Document Parsing (NEW)
pypdf>=3.0.0
python-docx>=0.8.11

# Database
sqlalchemy>=2.0.0
asyncpg
aiosqlite

# Utilities
python-dotenv
sentencepiece
# kss  # Unused and heavy

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart
